upgrade next and react
