+++ package.json
@@ -30,10 +30,10 @@
     "drizzle-orm": "^0.36.3",
     "input-otp": "^1.4.1",
     "lucide-react": "^0.460.0",
-    "next": "15.0.3",
+    "next": "^15.0.4",
     "next-themes": "^0.4.3",
-    "react": "19.0.0-rc-66855b96-20241106",
-    "react-dom": "19.0.0-rc-66855b96-20241106",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "react-hook-form": "^7.53.2",
     "react-icons": "^5.3.0",
     "resend": "^4.0.1",
