+++ src/lib/auth.ts
@@ -56,3 +56,5 @@ export const auth = betterAuth({
 		}),
 	],
 });
+
+export type Session = typeof auth.$Infer.Session;
